+++ R/extract_parameters_anova.R
@@ -42,6 +42,7 @@
   names(parameters) <- gsub("LR Chisq", "Chi2", names(parameters), fixed = TRUE)
   names(parameters) <- gsub("Chisq", "Chi2", names(parameters), fixed = TRUE)
   names(parameters) <- gsub("Chi.sq", "Chi2", names(parameters), fixed = TRUE)
+  names(parameters) <- gsub("Chi.Square", "Chi2", names(parameters), fixed = TRUE)
   names(parameters) <- gsub("Chi-Square", "Chi2", names(parameters), fixed = TRUE)
   # other
   names(parameters) <- gsub("logLik", "Log_Likelihood", names(parameters), fixed = TRUE)